Add files via upload
This game takes a total of one week（or less） to complete and was developed using the Unity engine. It is a playable game (rather than a demo) rather than a demonstration. Although most of the materials were sourced from the open-source marketplace, I designed and created the game UI myself.